update to project 2

+++ src/controller/animes.controller.js
@@ -1,31 +1,69 @@
 const animesService = require('../service/animes.service');
+const mongoose = require('mongoose');
 
-const findAllAnimesController = (req, res) => {
-  const animes = animesService.findAllAnimesService();
+const findAllAnimesController = async (req, res) => {
+  const animes = await animesService.findAllAnimesService();
+  if (animes.length == 0) {
+    return res
+      .status(404)
+      .send({ message: 'Não existe nem um anime cadastrado' });
+  }
   res.send(animes);
 };
 
-const findByIdAnimesController = (req, res) => {
-  const parametroId = Number(req.params.id);
-  const selectAnime = animesService.findByIdAnimesService(parametroId);
+const findByIdAnimesController = async (req, res) => {
+  const paramsId = req.params.id;
+
+  if (!mongoose.Types.ObjectId.isValid(paramsId)) {
+    return res.status(400).send({ message: 'Id inválido!' });
+  }
+  const selectAnime = await animesService.findByIdAnimesService(paramsId);
+  if (!selectAnime) {
+    return res.status(404).send({ message: 'Anime não encontrado!' });
+  }
+
   res.send(selectAnime);
 };
 
-const createAnimesController = (req, res) => {
-  const anime = req.body;
-  const newAnime = animesService.createAnimesService(anime);
-  res.send(newAnime);
+const deleteAnimeController = async (req, res) => {
+  const idParam = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    return res.status(400).send({ message: 'Id inválido!' });
+  }
+  await animesService.deleteAnimesService(idParam);
+
+  res.send({ message: 'Anime deletado!.' });
 };
 
-const deleteAnimesController = (req, res) => {
+const createAnimeController = async (req, res) => {
+  const anime = req.body;
+
+  if (!anime || !anime.name || !anime.year || !anime.autor) {
+    return res.status(400).send({ message: 'Envie todos os campos do anime!' });
+  }
+  const newAnime = await animesService.createAnimesService(anime);
+  res.status(201).send(newAnime);
+};
+
+const editAnimeController = async (req, res) => {
   const idParam = req.params.id;
-  animesService.deleteAnimesService(idParam);
-  res.send({ message: 'Anime successfully deleted' });
+  if (!mongoose.Types.ObjectId.isValid(idParam)) {
+    return res.status(400).send({ message: 'Id inválido!' });
+  }
+
+  const animeEdit = req.body;
+
+  if (!animeEdit || !animeEdit.name || !animeEdit.year || !animeEdit.autor) {
+    return res.status(400).send({ message: 'Envie todos os campos do anime!' });
+  }
+  const editedAnime = await animesService.editAnimesService(idParam, animeEdit);
+  res.send(editedAnime);
 };
 
 module.exports = {
   findAllAnimesController,
   findByIdAnimesController,
-  createAnimesController,
-  deleteAnimesController,
+  deleteAnimeController,
+  createAnimeController,
+  editAnimeController,
 };
